feat(ci): add  option to ignore specific checks
If you want to ignore a single check in your CI today you have to
specify all check names _except_ the one you want to ignore in the
`--check=<check>` flags. This can create a burden to ensure your
arguments stay up to date with any CI check changes you make.

This change adds a different option to specify `--exclude=<check>`. So
if you add or change other checks, wait-for-github still continues to
work as expected. This argument is less strict than `--check=<check>`
meaning if you specify checks specifically, there should be no need to
exclude any of the checks, so the argument is ignored in that case.